requirements.txt
PyYAML>=6.0
openai>=1.0.0